+++ tools/developer/ltsconvertsymbolic/ltsconvertsymbolic.cpp
@@ -7,6 +7,8 @@
 // http://www.boost.org/LICENSE_1_0.txt)
 //
 
+#include <cstddef>
+
 #include "mcrl2/utilities/input_output_tool.h"
 
 #include "mcrl2/lps/symbolic_lts_io.h"
@@ -98,7 +100,7 @@ class ltsconvert_tool : public input_output_tool
 
   // Lace options
   std::size_t lace_n_workers = 1;
-  std::size_t lace_dqsize = 1024 * 1024 * 4; // set large default
+  std::size_t lace_dqsize = static_cast<std::size_t>(1024 * 1024 * 4); // set large default
   std::size_t lace_stacksize = 0;            // use default
 
   // Sylvan options
